requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
greenlet==2.0.1
playwright==1.30.0
pyee==9.0.4
python-dotenv==0.21.1
soupsieve==2.3.2.post1
typing-extensions==4.4.0
